convert metrics to ts
